refactor(admin): Q9.g — extract admin/users.rs (GAR-474)
Moves setup/users/danger-zone handlers (lines 174–543, ~370 LOC) from
admin/handlers.rs into a new focused module admin/users.rs.
Re-exports in handlers.rs preserve all routes.rs call sites unchanged.
handlers.rs shrinks from 2103 to 1738 LOC.

GAR-474 / Epic GAR-430 (Q9 admin/handlers.rs refactor series, slice Q9.g).
