+++ crates/garraia-gateway/src/admin/mod.rs
@@ -9,3 +9,4 @@ pub mod rbac;
 pub mod routes;
 pub mod shared;
 pub mod store;
+pub mod users;
